requirements.txt
matplotlib==3.9.0
mysql_connector_repackaged==0.3.1
numpy==2.0.0
pandas==1.5.3
Requests==2.32.3
